Crate webp_animation

Overview

This crate provides a high-level Rust wrapper for decoding and encoding WebP animations. Underlying WebP format processing is handled by C-based libwebp library by Google, which is interfaced through Rust libwebp-sys2 crate

Usage

Have a look at Decoder and Encoder for use-case specific examples.

Modules

prelude

Structs

Decoder

A decoder for webp animation data

DecoderIterator

An iterator that produces decoded Frame’s from webp data

DecoderOptions

An options struct for Decoder

Encoder

An encoder for creating webp animation

EncoderOptions

An options struct for Encoder instance

EncodingConfig

Encoding configuration. Can be set for Encoder globally or per frame

Frame

An animation frame containing data and metadata produced by Decoder

LossyEncodingConfig

Parameters related to lossy compression only

WebPData

A safe wrapper for WebP bytedata. Consider as &[u8] (implements Deref)

Enums

ColorMode

Color Mode that configures the output type of Decoder Frame’s

EncodingType

Encoding type

Error

Error type produced by webp_animation code